+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Fixed
 - Throwing a non-error from tests breaks various reporters
 - Replaced deprecated assert.equal with assert.strictEqual
+- SpecReporter colours could not be disabled
 
 ### Added
 - Discover tripitaka badge
+- SpecReporter tests